sbcproject/requirement.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.1.12
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.4.2
idna==3.4
pip==23.1.2
pymongo==3.12.3
pytz==2023.3.post1
requests==2.31.0
setuptools==67.8.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.0.6
wheel==0.40.0